+++ Changelog.md
@@ -14,6 +14,7 @@
 * Fix medium and small avatar URLs when using Camo [#5883](https://github.com/diaspora/diaspora/pull/5883)
 * Improve output of script/server [#5885](https://github.com/diaspora/diaspora/pull/5885)
 * Fix CSS for bold links [#5887](https://github.com/diaspora/diaspora/pull/5887)
+* Correctly handle IE8 in the chrome frame middleware [#5878](https://github.com/diaspora/diaspora/pull/5878)
 
 ## Features
 * Hide post title of limited post in comment notification email [#5843](https://github.com/diaspora/diaspora/pull/5843)